# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Telecom Call Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User management
  - name: calls
    description: Call management
  - name: payments
    description: Payment operations
  - name: numbers
    description: Phone number management
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List call history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
    post:
      operationId: makeCall
      tags: [calls]
      summary: Initiate a new call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeCallRequest"
      responses:
        "200":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallRecord"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /calls/{callId}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get a specific call record
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallRecord"

  /numbers:
    get:
      operationId: listNumbers
      tags: [numbers]
      summary: List all phone numbers with ownership status
      responses:
        "200":
          description: Phone numbers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberListResponse"
        "403":
          description: Subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /numbers/select:
    post:
      operationId: selectNumber
      tags: [numbers]
      summary: Select/claim a free phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectNumberRequest"
      responses:
        "200":
          description: Number assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberActionResponse"
        "400":
          description: Limit reached or taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /numbers/change:
    post:
      operationId: changeNumber
      tags: [numbers]
      summary: Initiate number change (R100 fee via PayFast)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeNumberRequest"
      responses:
        "200":
          description: PayFast payment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayFastPaymentData"

  /payments/subscribe:
    post:
      operationId: initiateSubscription
      tags: [payments]
      summary: Initiate a subscription payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: PayFast payment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayFastPaymentData"

  /payments/webhook:
    post:
      operationId: payfastWebhook
      tags: [payments]
      summary: PayFast ITN webhook
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Webhook processed

  /payments/history:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment history
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"

  /credits/topup:
    post:
      operationId: topUpCredits
      tags: [payments]
      summary: Top up wallet coins via PayFast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpRequest"
      responses:
        "200":
          description: PayFast payment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayFastPaymentData"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"

  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Admin - get user details
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"

  /admin/users/{userId}/adjust-credit:
    post:
      operationId: adminAdjustCredit
      tags: [admin]
      summary: Admin - manually adjust user coins
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCreditRequest"
      responses:
        "200":
          description: Coins adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin - get platform statistics
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/calls:
    get:
      operationId: adminListCalls
      tags: [admin]
      summary: Admin - list all calls
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        profileImage:
          type: string
        isAdmin:
          type: boolean
      required: [id, username, isAdmin]

    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        profileImage:
          type: string
        email:
          type: string
        coins:
          type: number
          description: Wallet coin balance (1 coin = R0.90)
        subscriptionStatus:
          type: string
          enum: [active, inactive, pending]
        subscriptionPlan:
          type: string
          enum: [basic, pro]
        subscriptionExpiresAt:
          type: string
          format: date-time
          nullable: true
        lastPaymentDate:
          type: string
          format: date-time
          nullable: true
        nextPaymentDate:
          type: string
          format: date-time
          nullable: true
        totalCallsUsed:
          type: integer
        totalCoinsUsed:
          type: number
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - coins
        - subscriptionStatus
        - subscriptionPlan
        - totalCallsUsed
        - totalCoinsUsed
        - isAdmin
        - createdAt

    PhoneNumber:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        status:
          type: string
          enum: [owned, taken, free]
        userId:
          type: string
          nullable: true
      required: [id, number, status]

    OwnedNumber:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
      required: [id, number]

    NumberListResponse:
      type: object
      properties:
        numbers:
          type: array
          items:
            $ref: "#/components/schemas/PhoneNumber"
        myNumbers:
          type: array
          items:
            $ref: "#/components/schemas/OwnedNumber"
        maxNumbers:
          type: integer
        plan:
          type: string
      required: [numbers, myNumbers, maxNumbers, plan]

    SelectNumberRequest:
      type: object
      properties:
        numberId:
          type: string
      required: [numberId]

    ChangeNumberRequest:
      type: object
      properties:
        oldNumberId:
          type: string
        newNumberId:
          type: string
      required: [oldNumberId, newNumberId]

    NumberActionResponse:
      type: object
      properties:
        message:
          type: string
        number:
          $ref: "#/components/schemas/OwnedNumber"
      required: [message, number]

    CallRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        callerNumber:
          type: string
        recipientNumber:
          type: string
        status:
          type: string
          enum: [initiated, ringing, in-progress, completed, failed, busy, no-answer]
        duration:
          type: integer
        cost:
          type: number
          description: Coins used
        telnyxCallId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
      required: [id, userId, recipientNumber, status, duration, cost, createdAt]

    CallListResponse:
      type: object
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/CallRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [calls, total, page, limit, totalPages]

    MakeCallRequest:
      type: object
      properties:
        recipientNumber:
          type: string
        notes:
          type: string
      required: [recipientNumber]

    SubscribeRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [basic, pro]
      required: [plan]

    PayFastPaymentData:
      type: object
      properties:
        paymentUrl:
          type: string
        amount:
          type: string
        itemName:
          type: string
        paymentId:
          type: string
        coinsAdded:
          type: integer
        formFields:
          type: object
          additionalProperties:
            type: string
      required: [paymentUrl, amount, itemName, paymentId, formFields]

    TopUpRequest:
      type: object
      properties:
        amount:
          type: number
      required: [amount]

    PaymentRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        coinsAdded:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        paymentType:
          type: string
          enum: [subscription, topup, number_change]
        subscriptionPlan:
          type: string
          nullable: true
        payfastPaymentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, amount, coinsAdded, status, paymentType, createdAt]

    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
        total:
          type: integer
      required: [payments, total]

    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, limit, totalPages]

    AdminUserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        recentCalls:
          type: array
          items:
            $ref: "#/components/schemas/CallRecord"
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
      required: [user, recentCalls, recentPayments]

    AdjustCreditRequest:
      type: object
      properties:
        amount:
          type: number
        reason:
          type: string
      required: [amount, reason]

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeSubscriptions:
          type: integer
        totalCalls:
          type: integer
        totalCallMinutes:
          type: integer
        totalRevenue:
          type: number
        callsToday:
          type: integer
        newUsersThisMonth:
          type: integer
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
      required:
        - totalUsers
        - activeSubscriptions
        - totalCalls
        - totalCallMinutes
        - totalRevenue
        - callsToday
        - newUsersThisMonth
        - recentPayments