# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecom Call Manager API
 * OpenAPI spec version: 0.2.0
 */
import type { PaymentRecord } from "./paymentRecord";

export interface AdminStats {
  totalUsers: number;
  activeSubscriptions: number;
  totalCalls: number;
  totalCallMinutes: number;
  totalRevenue: number;
  callsToday: number;
  newUsersThisMonth: number;
  recentPayments: PaymentRecord[];
}